+++ modules/overrides/userconfigurable/client/limits.go
@@ -1,6 +1,7 @@
 package client
 
 import (
+	filterconfig "github.com/grafana/tempo/pkg/spanfilter/config"
 	"github.com/grafana/tempo/pkg/util/listtomap"
 )
 
@@ -99,8 +100,9 @@ func (l *LimitsMetricsGeneratorProcessorServiceGraphs) GetPeerAttributes() ([]st
 }
 
 type LimitsMetricsGeneratorProcessorSpanMetrics struct {
-	Dimensions       *[]string `json:"dimensions,omitempty"`
-	EnableTargetInfo *bool     `json:"enable_target_info,omitempty"`
+	Dimensions       *[]string                    `json:"dimensions,omitempty"`
+	EnableTargetInfo *bool                        `json:"enable_target_info,omitempty"`
+	FilterPolicies   *[]filterconfig.FilterPolicy `json:"filter_policies,omitempty"`
 }
 
 func (l *LimitsMetricsGeneratorProcessorSpanMetrics) GetDimensions() ([]string, bool) {
@@ -115,4 +117,11 @@ func (l *LimitsMetricsGeneratorProcessorSpanMetrics) GetEnableTargetInfo() (bool
 		return *l.EnableTargetInfo, true
 	}
 	return false, false
+}
+
+func (l *LimitsMetricsGeneratorProcessorSpanMetrics) GetFilterPolicies() ([]filterconfig.FilterPolicy, bool) {
+	if l != nil && l.FilterPolicies != nil {
+		return *l.FilterPolicies, true
+	}
+	return nil, true
 }